refactor(cli): extract macro/profile command subsystem module (#166)
